Check for only known references:
If you now try to match a key's value in a parameter against an
unknown reference (currently only code and regexp reference are
recognized), it will crash, because it won't be able to match it
correctly.

In order to match it correctly, it will need to traverse all the
tree and compare the values instead of the reference, because the
reference will definitely not be the same (param reference vs.
rules reference).

So, instead of going all fucking apeshit, we're just not allowing
it. You'll thank us later! :)